fix harmless warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5257 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/qsound.c
@@ -184,7 +184,7 @@ int qsound_sh_start(const struct MachineSound *msound)
 	const struct QSound_interface *intf = msound->sound_interface;
 
 	qsound_sample_rom = (uint8_t *)memory_region(intf->region);
-	qsound_sample_rom_length = memory_region_length(intf->region);
+	qsound_sample_rom_length = (unsigned int)memory_region_length(intf->region);
 	qsound_sample_rom_mask = pow2_mask(qsound_sample_rom_length);
 
 	m_vgm_idx = vgm_open(VGMC_QSOUND, intf->clock);
